Correctly filter style rules in overrides
